claims and custom error messages in jwt

+++ app/resources/item.py
@@ -1,8 +1,10 @@
 
-import uuid
+
 from flask import request
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
+from flask_jwt_extended import jwt_required,get_jwt  # protect endpoint using jwt/accesstoken
+
 from app.schemas import ItemSchema, ItemUpdateSchema
 from app.models.item import ItemModel
 from app.models.store import StoreModel
@@ -17,12 +19,14 @@ blp = Blueprint("Items", __name__, description="Opertation on items")
 @blp.route("/item/<int:item_id>")
 class Item(MethodView):
 
+    @jwt_required()
     @blp.response(200,ItemSchema)
     def get(self, item_id):
 
         item = ItemModel.query.get_or_404(item_id)
         return item
             
+    @jwt_required()
     @blp.arguments(ItemUpdateSchema)
     @blp.response(200, ItemSchema)
     def put(self, item_data, item_id):
@@ -41,7 +45,13 @@ class Item(MethodView):
 
         return item
     
+    @jwt_required()
     def delete(self,item_id):
+
+        jwt = get_jwt() # get jwt claims
+        if not jwt.get("is_admin"):   # get specific data from jwt claim
+            abort(401, message="admin privilege required")
+        
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
@@ -51,10 +61,13 @@ class Item(MethodView):
 @blp.route("/item")
 class ItemList(MethodView):
 
+    @jwt_required()
     @blp.response(200, ItemSchema(many=True))   # converts dictionary to a list
     def get(self):
         return ItemModel.query.all()
 
+
+    @jwt_required() # proteect endpoint by requring a jwt
     # to apply schemna
     @blp.arguments(ItemSchema)
     @blp.response(200, ItemSchema)
